fix: using forge format

+++ test/DappsManager.t.sol
@@ -10,11 +10,7 @@ contract DappsManagerTest is Test {
     DappRank public drnkToken;
 
     address nonAdmin = address(0x20);
-    address[] testUsers = [ address(0x1),
-        address(0x2),
-        address(0x3),
-        address(0x4),
-        address(0x5)];
+    address[] testUsers = [address(0x1), address(0x2), address(0x3), address(0x4), address(0x5)];
 
     uint256 listingFee = 100e16;
     uint256 burningFee = 1000; // 10%
@@ -22,12 +18,7 @@ contract DappsManagerTest is Test {
     uint256 bonus = 1000e18;
 
     function setUp() public {
-        dappsMgr = new DappsManager(
-            listingFee,
-            daoFee,
-            burningFee,
-            bonus
-        );
+        dappsMgr = new DappsManager(listingFee, daoFee, burningFee, bonus);
         // Attach deployed token
         drnkToken = DappRank(address(dappsMgr.drnk()));
     }
@@ -42,7 +33,7 @@ contract DappsManagerTest is Test {
         string memory cid = "QmTestCID";
 
         // ToDo, add custom Error
-        vm.expectRevert( "Error: listing fee uncovered");
+        vm.expectRevert("Error: listing fee uncovered");
         dappsMgr.registerDapp{value: 0}(dappName, cid);
     }
 
@@ -84,7 +75,7 @@ contract DappsManagerTest is Test {
         // Retrive Dapp
         string memory retCID;
         bytes32 status;
-        (retCID, , , , , , , status) = dappsMgr.getDappInfo(dappName);
+        (retCID,,,,,,, status) = dappsMgr.getDappInfo(dappName);
         //console2.log("retrived status: ");
         //console2.logBytes32(status);
         //console2.log("retrived cid: ", retCID);
@@ -92,17 +83,17 @@ contract DappsManagerTest is Test {
         assertEq(cid, retCID);
         // Then Dapp gets approved
         dappsMgr.approveDapp(dappName);
-        (retCID, , , , , , , status) = dappsMgr.getDappInfo(dappName);
+        (retCID,,,,,,, status) = dappsMgr.getDappInfo(dappName);
         assertEq(status, bytes32("Active"));
         assertEq(cid, retCID);
         // After the Dapp gets banned
         dappsMgr.banDapp(dappName);
-        (retCID, , , , , , , status) = dappsMgr.getDappInfo(dappName);
+        (retCID,,,,,,, status) = dappsMgr.getDappInfo(dappName);
         assertEq(status, bytes32("Banned"));
         assertEq(cid, retCID);
         // Then Dapp gets approved Again
         dappsMgr.approveDapp(dappName);
-        (retCID, , , , , , , status) = dappsMgr.getDappInfo(dappName);
+        (retCID,,,,,,, status) = dappsMgr.getDappInfo(dappName);
         assertEq(status, bytes32("Active"));
         assertEq(cid, retCID);
         // Remove the Dapp
@@ -120,7 +111,7 @@ contract DappsManagerTest is Test {
         // Retrive Dapp
         string memory retCID;
         bytes32 status;
-        (retCID, , , , , , , status) = dappsMgr.getDappInfo(dappName);
+        (retCID,,,,,,, status) = dappsMgr.getDappInfo(dappName);
         assertEq(status, bytes32("Submitted"));
         assertEq(cid, retCID);
         // Remove Dapp knowing the index...
@@ -129,44 +120,32 @@ contract DappsManagerTest is Test {
     }
 
     function testAddandRemoveDappMany() external {
-        bytes32[5] memory dappNames = [
-            bytes32("dapp1"),
-            bytes32("dapp2"),
-            bytes32("dapp3"),
-            bytes32("dapp4"),
-            bytes32("dapp5")
-        ];
-        string[5] memory cids = [
-            "QmNewCID1",
-            "QmNewCID2",
-            "QmNewCID3",
-            "QmNewCID4",
-            "QmNewCID5"
-        ];
+        bytes32[5] memory dappNames =
+            [bytes32("dapp1"), bytes32("dapp2"), bytes32("dapp3"), bytes32("dapp4"), bytes32("dapp5")];
+        string[5] memory cids = ["QmNewCID1", "QmNewCID2", "QmNewCID3", "QmNewCID4", "QmNewCID5"];
         // Retrive Dapp
         string memory retCID;
         bytes32 status;
 
-        for(uint i; i < dappNames.length; i++) {
+        for (uint256 i; i < dappNames.length; i++) {
             dappsMgr.registerDapp{value: listingFee}(dappNames[i], cids[i]);
         }
 
-        for(uint i; i < dappNames.length; i++) {
-            (retCID, , , , , , , status) = dappsMgr.getDappInfo(dappNames[i]);
+        for (uint256 i; i < dappNames.length; i++) {
+            (retCID,,,,,,, status) = dappsMgr.getDappInfo(dappNames[i]);
             assertEq(status, bytes32("Submitted"));
             assertEq(retCID, cids[i]);
         }
 
-
         // remove all (ToDo, fix math looks dirty)
         bytes32[] memory dappN = dappsMgr.getAllDappNames();
-        for(uint i = dappN.length; i > 0; i--) {
+        for (uint256 i = dappN.length; i > 0; i--) {
             //console2.log(i-1);
-            console2.log(dappsMgr.DappNameExists(dappNames[i-1]));
-            dappsMgr.removeDapp(i-1, dappNames[i-1]);
-            console2.logBytes32(dappNames[i-1]);
-            console2.log("removed dappname:",string(abi.encodePacked(dappNames[i-1])));
-            console2.log(dappsMgr.DappNameExists(dappNames[i-1]));
+            console2.log(dappsMgr.DappNameExists(dappNames[i - 1]));
+            dappsMgr.removeDapp(i - 1, dappNames[i - 1]);
+            console2.logBytes32(dappNames[i - 1]);
+            console2.log("removed dappname:", string(abi.encodePacked(dappNames[i - 1])));
+            console2.log(dappsMgr.DappNameExists(dappNames[i - 1]));
             dappN = dappsMgr.getAllDappNames();
         }
         assertEq(dappsMgr.getAllDappNames().length, 0);
@@ -174,12 +153,12 @@ contract DappsManagerTest is Test {
 
     function testBurnTestUsersTokens() public {
         dappsMgr.demoAirdrop(testUsers);
-        for(uint i; i < testUsers.length; i++) {
+        for (uint256 i; i < testUsers.length; i++) {
             assertEq(drnkToken.balanceOf(testUsers[i]), bonus);
         }
 
         // test burning from ERC20
-        for(uint i; i < testUsers.length; i++) {
+        for (uint256 i; i < testUsers.length; i++) {
             vm.prank(testUsers[i]);
             drnkToken.burn(bonus);
             assertEq(drnkToken.balanceOf(testUsers[i]), 0);
@@ -187,12 +166,12 @@ contract DappsManagerTest is Test {
 
         // TEST delegated burn
         dappsMgr.demoAirdrop(testUsers);
-        for(uint i; i < testUsers.length; i++) {
+        for (uint256 i; i < testUsers.length; i++) {
             assertEq(drnkToken.balanceOf(testUsers[i]), bonus);
         }
 
         // Allow drnkTokens from testUsers to dappsMgr
-        for(uint i; i < testUsers.length; i++) {
+        for (uint256 i; i < testUsers.length; i++) {
             vm.prank(testUsers[i]);
             drnkToken.approve(address(dappsMgr), bonus);
             uint256 allowanceAmount = drnkToken.allowance(testUsers[i], address(dappsMgr));
@@ -200,65 +179,53 @@ contract DappsManagerTest is Test {
         }
 
         // test burning from ERC20
-        for(uint i; i < testUsers.length; i++) {
+        for (uint256 i; i < testUsers.length; i++) {
             vm.prank(testUsers[i]);
             dappsMgr.burn(bonus);
             assertEq(drnkToken.balanceOf(testUsers[i]), 0);
         }
-
     }
 
     function testVote4Dapp() public {
         // cash in the demo users
         dappsMgr.demoAirdrop(testUsers);
-        for(uint i; i < testUsers.length; i++) {
+        for (uint256 i; i < testUsers.length; i++) {
             assertEq(drnkToken.balanceOf(testUsers[i]), bonus);
         }
         // simulated dapps
-        bytes32[5] memory dappNames = [
-            bytes32("dapp1"),
-            bytes32("dapp2"),
-            bytes32("dapp3"),
-            bytes32("dapp4"),
-            bytes32("dapp5")
-        ];
-        string[5] memory cids = [
-            "QmNewCID1",
-            "QmNewCID2",
-            "QmNewCID3",
-            "QmNewCID4",
-            "QmNewCID5"
-        ];
+        bytes32[5] memory dappNames =
+            [bytes32("dapp1"), bytes32("dapp2"), bytes32("dapp3"), bytes32("dapp4"), bytes32("dapp5")];
+        string[5] memory cids = ["QmNewCID1", "QmNewCID2", "QmNewCID3", "QmNewCID4", "QmNewCID5"];
         // Retrive Dapp
         string memory retCID;
         bytes32 status;
 
-        for(uint i; i < dappNames.length; i++) {
+        for (uint256 i; i < dappNames.length; i++) {
             vm.prank(testUsers[i]);
             vm.deal(testUsers[i], 10 ether);
             dappsMgr.registerDapp{value: listingFee}(dappNames[i], cids[i]);
         }
 
-        for(uint i; i < dappNames.length; i++) {
-            (retCID, , , , , , , status) = dappsMgr.getDappInfo(dappNames[i]);
+        for (uint256 i; i < dappNames.length; i++) {
+            (retCID,,,,,,, status) = dappsMgr.getDappInfo(dappNames[i]);
             assertEq(status, bytes32("Submitted"));
             assertEq(retCID, cids[i]);
         }
-        for(uint i; i < dappNames.length; i++) {
+        for (uint256 i; i < dappNames.length; i++) {
             dappsMgr.approveDapp(dappNames[i]);
         }
-        for(uint i; i < dappNames.length; i++) {
-            (retCID, , , , , , , status) = dappsMgr.getDappInfo(dappNames[i]);
+        for (uint256 i; i < dappNames.length; i++) {
+            (retCID,,,,,,, status) = dappsMgr.getDappInfo(dappNames[i]);
             assertEq(status, bytes32("Active"));
             assertEq(retCID, cids[i]);
         }
 
         // Voting Process
 
-        uint vote_amount = 100e18;
+        uint256 vote_amount = 100e18;
 
         //// allow daggTokens from testUsers to dappsMgr
-        for(uint i; i < testUsers.length; i++) {
+        for (uint256 i; i < testUsers.length; i++) {
             vm.prank(testUsers[i]);
             drnkToken.approve(address(dappsMgr), (vote_amount * 2 * i) + vote_amount); // times I for make a dyanmic PoC
             uint256 allowanceAmount = drnkToken.allowance(testUsers[i], address(dappsMgr));
@@ -266,11 +233,11 @@ contract DappsManagerTest is Test {
         }
 
         //// test Voting
-        for(uint i; i < testUsers.length; i++) {
+        for (uint256 i; i < testUsers.length; i++) {
             // we only vote for the first dapp in this demo.
             // simulate vote with the dyamic amount where rates are Sumi(i*10+1)=1,25,49,73,97.
             vm.prank(testUsers[i]);
-            dappsMgr.voteDapp(dappNames[0], (vote_amount * 2 * i) + vote_amount, i*24 +1);
+            dappsMgr.voteDapp(dappNames[0], (vote_amount * 2 * i) + vote_amount, i * 24 + 1);
             // ToDo for better debug the access with solidity gimnastics
             //console2.log("-----------------");
             //console2.log("User", testUsers[i]);
@@ -286,7 +253,7 @@ contract DappsManagerTest is Test {
         uint256 balance;
         uint256 burned;
         address owner;
-        (, rate, weight_votes_sum, weight_total_sum, balance, burned, owner, ) = dappsMgr.getDappInfo(dappNames[0]);
+        (, rate, weight_votes_sum, weight_total_sum, balance, burned, owner,) = dappsMgr.getDappInfo(dappNames[0]);
         assertEq(rate, 60);
         assertEq(weight_votes_sum, 6380084467978);
         assertEq(weight_total_sum, 106138700962);
@@ -320,7 +287,7 @@ contract DappsManagerTest is Test {
     }
 
     function testTopUp() public {
-        uint topUpAmount = 1 ether;
+        uint256 topUpAmount = 1 ether;
         assertEq(address(testUsers[0]).balance, 0 ether);
         vm.prank(testUsers[0]);
         vm.deal(testUsers[0], 1.1 ether);
@@ -334,5 +301,4 @@ contract DappsManagerTest is Test {
     //function testAddandRemovefans() external {}
     //function testDappsExpire() public {}
     //function testFanLifeTime() public {}
-
 }
